Move the parts of CollectionExtensions which do not need anything local (types, Random) into Tumult...
