+++ SpaceWarDiscordApp/CollectionExtensions.cs
@@ -1,102 +1,45 @@
 ﻿using System.Collections;
 using SpaceWarDiscordApp.GameLogic.Techs;
 
-namespace SpaceWarDiscordApp
+namespace SpaceWarDiscordApp;
+
+internal static class CollectionExtensions
 {
-    internal static class CollectionExtensions
+    public static T Random<T>(this IReadOnlyList<T> list) => list[Program.Random.Next(list.Count)];
+    public static int RandomIndex<T>(this IReadOnlyList<T> list) => Program.Random.Next(list.Count);
+    public static T Random<T>(this IReadOnlyList<T> list, Func<T, float> weightSelector)
     {
-        public static IEnumerable<int> Between(int start, int end) => Enumerable.Range(start, end - start + 1);
-
-        public static IEnumerable<(T first, T second)> Combinations<T>(this IReadOnlyList<T> list)
+        var totalWeight = list.Sum(weightSelector);
+        var selected = Program.Random.NextDouble() * totalWeight;
+        var skipped = 0f;
+        return list.SkipWhile(x => (skipped += weightSelector(x)) <= selected).FirstOrDefault() ?? list[^1];
+    }
+    public static IEnumerable<T> RandomUnique<T>(this IReadOnlyList<T> list, int count) => list.Shuffled().Take(count);
+    
+    public static IList<T> Shuffled<T>(this IEnumerable<T> list)
+    {
+        var copy = new List<T>(list);
+        copy.Shuffle();
+        return copy;
+    }
+    
+    /// <summary>
+    /// Randomly reorders a list in place
+    /// </summary>
+    public static void Shuffle<T>(this IList<T> list)
+    {
+        /* To shuffle an array a of n elements (indices 0..n-1):
+        for i from n−1 downto 1 do
+                j ← random integer such that 0 ≤ j ≤ i
+                exchange a[j] and a[i]*/
+    
+        for (int i = list.Count - 1; i > 0; i--)
         {
-            for (var i = 0; i < list.Count; i++)
-            {
-                for (var j = i + 1; j < list.Count; j++)
-                {
-                    yield return (list[i], list[j]);
-                }
-            }
-        }
-        
-        public static T Random<T>(this IReadOnlyList<T> list) => list[Program.Random.Next(list.Count)];
-        public static int RandomIndex<T>(this IReadOnlyList<T> list) => Program.Random.Next(list.Count);
-        public static T Random<T>(this IReadOnlyList<T> list, Func<T, float> weightSelector)
-        {
-            var totalWeight = list.Sum(weightSelector);
-            var selected = Program.Random.NextDouble() * totalWeight;
-            var skipped = 0f;
-            return list.SkipWhile(x => (skipped += weightSelector(x)) <= selected).FirstOrDefault() ?? list[^1];
-        }
-        public static IEnumerable<T> RandomUnique<T>(this IReadOnlyList<T> list, int count) => list.Shuffled().Take(count);
-
-        public static IList<T> Shuffled<T>(this IEnumerable<T> list)
-        {
-            var copy = new List<T>(list);
-            copy.Shuffle();
-            return copy;
-        }
-
-        /// <summary>
-        /// Randomly reorders a list in place
-        /// </summary>
-        public static void Shuffle<T>(this IList<T> list)
-        {
-            /* To shuffle an array a of n elements (indices 0..n-1):
-            for i from n−1 downto 1 do
-                    j ← random integer such that 0 ≤ j ≤ i
-                    exchange a[j] and a[i]*/
-
-            for (int i = list.Count - 1; i > 0; i--)
-            {
-                var j = Program.Random.Next(0, i + 1);
-                (list[i], list[j]) = (list[j], list[i]);
-            }
-        }
-
-        public static IEnumerable<Tech> ToTechsById(this IEnumerable<string> techIds) => techIds.Select(x => Tech.TechsById[x]);
-        public static IEnumerable<Tech?> ToTechsByIdNullable(this IEnumerable<string?> techIds) => techIds.Select(x => x == null ? null : Tech.TechsById[x]);
-
-        public static IEnumerable<int> Indices(this ICollection collection) => Enumerable.Range(0, collection.Count);
-
-        public static IEnumerable<(T item, int index)> ZipWithIndices<T>(this IEnumerable<T> collection) => collection.Select((x, i) => (x, i));
-        
-        public static IEnumerable<T> Except<T>(this IEnumerable<T> enumerable, T item) => enumerable.Where(x => !EqualityComparer<T>.Default.Equals(x, item));
-        
-        public static IEnumerable<T> WhereNonNull<T>(this IEnumerable<T?> enumerable) => enumerable.Where(x => x != null)!;
-        public static IAsyncEnumerable<T> WhereNonNull<T>(this IAsyncEnumerable<T?> enumerable) => enumerable.Where(x => x != null)!;
-        public static IEnumerable<T1> WhereNonNull<T1, T2>(this IEnumerable<T1> enumerable, Func<T1, T2?> extractor) => enumerable.Where(x => extractor(x) != null);
-
-        /// <summary>
-        /// Combines two sequences into a single sequence of tuples. If one sequence is shorter, it is padded with
-        /// default values to match the length other sequence
-        /// </summary>
-        public static IEnumerable<(T1?, T2?)> ZipLongest<T1, T2>(this IEnumerable<T1> enumerable,
-            IEnumerable<T2> enumerable2)
-        {
-            using var enumerator1 = enumerable.GetEnumerator();
-            using var enumerator2 = enumerable2.GetEnumerator();
-
-            var hasLeft = enumerator1.MoveNext();
-            var hasRight = enumerator2.MoveNext();
-
-            while (hasLeft || hasRight)
-            {
-                if (hasLeft && hasRight)
-                {
-                    yield return (enumerator1.Current, enumerator2.Current);
-                }
-                else if (hasLeft)
-                {
-                    yield return (enumerator1.Current, default);
-                }
-                else if (hasRight)
-                {
-                    yield return (default, enumerator2.Current);
-                }
-
-                hasLeft = enumerator1.MoveNext();
-                hasRight = enumerator2.MoveNext();
-            }
+            var j = Program.Random.Next(0, i + 1);
+            (list[i], list[j]) = (list[j], list[i]);
         }
     }
+
+    public static IEnumerable<Tech> ToTechsById(this IEnumerable<string> techIds) => techIds.Select(x => Tech.TechsById[x]);
+    public static IEnumerable<Tech?> ToTechsByIdNullable(this IEnumerable<string?> techIds) => techIds.Select(x => x == null ? null : Tech.TechsById[x]);
 }
