+++ SpaceWarDiscordApp/ImageGeneration/BoardImageGenerator.cs
@@ -1,18 +1,22 @@
 using System.Diagnostics.Contracts;
 using System.Numerics;
 using System.Text;
+
 using SixLabors.Fonts;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Drawing;
 using SixLabors.ImageSharp.Drawing.Processing;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
+using Path = SixLabors.ImageSharp.Drawing.Path;
+
 using SpaceWarDiscordApp.Database;
 using SpaceWarDiscordApp.Database.EventRecords;
 using SpaceWarDiscordApp.GameLogic;
 using SpaceWarDiscordApp.GameLogic.Operations;
 using SpaceWarDiscordApp.GameLogic.Techs;
-using Path = SixLabors.ImageSharp.Drawing.Path;
+
+using Tumult;
 
 // ImageSharp produces a lot of this due to the Mutate() etc. methods taking delegates
 // but it's fine as they do in fact run immediately and not access captured variables later on
